Chat (#10)
* * Added editing ticket
* Added rich text box
* Modified popup for editing tickets
* Added loading from an existing ticket to edit it
* Added modifying ticket status by drag and drop
* Added update burndown right after changing ticket status

* * Added project tickets counter for database
* Done database migration
* Update ticket status on drag and drop can now update without reload on
  ticket details while editing
* Delete functionality
* Confirm deletion functionality
* Fixed bug with quill rich text editor
* More CSS
* Fixed column layout formatting on project board

* added LogInfoModel and data class, not yet updated database

* added functionality to store logged user actions into database, database not migrated currently

* Notifications renamed, restructured, database updated and ready to integrate with frontend to display

* fetching notifications and loading in bottom of projectboard now functional

* added formatted statuses for messages, cleaned up

* changes

* added dynamic coloring to priority tags in tickets

---------

+++ Models/Projects/ProjectBoardModel.cs
@@ -3,7 +3,6 @@ using CodeBuggy.Data;
 using CodeBuggy.Helpers;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Build.Evaluation;
 using Microsoft.CodeAnalysis;
 using Microsoft.EntityFrameworkCore;
 using System.ComponentModel.DataAnnotations;
@@ -18,9 +17,10 @@ public class ProjectBoardModel
     private readonly AppDbContext _context;
     private readonly UserManager<AppUser> _userManager;
     private readonly BurndownModel _burndownModel;
+    private readonly NotificationModel _notificationModel;
 
     public List<Ticket> Tickets { get; set; }
-    
+
     [BindProperty]
     public InputModel Input { get; set; }
 
@@ -49,6 +49,7 @@ public class ProjectBoardModel
         _context = context;
         _userManager = userManager;
         _burndownModel = new BurndownModel();
+        _notificationModel = new NotificationModel();
     }
 
     public async Task<OperationResult> AddTicketToProject(ClaimsPrincipal user, InputModel input, int projectId)
@@ -69,8 +70,8 @@ public class ProjectBoardModel
             return new OperationResult { Success = false, Message = "User is not authenticated" };
         }
 
-        var owner = _userManager?.GetUserAsync(user)?.Result?.FirstName + " " + _userManager?.GetUserAsync(user)?.Result?.LastName;
-        if (owner == null)
+        var username = _userManager?.GetUserAsync(user)?.Result?.FirstName + " " + _userManager?.GetUserAsync(user)?.Result?.LastName;
+        if (username == null)
         {
             return new OperationResult { Success = false, Message = "User is not authenticated" };
         }
@@ -82,7 +83,7 @@ public class ProjectBoardModel
             Title = input.TicketTitle,
             Priority = input.TicketPriorityValue >= TicketPriority.None ? input.TicketPriorityValue : TicketPriority.None,
             Status = input.TicketStatusValue >= TicketStatus.ToDo ? input.TicketStatusValue : TicketStatus.ToDo,
-            CreatedBy = owner,
+            CreatedBy = username,
             CreationDate = DateTime.UtcNow,
             ResolvedBy = "",
             Description = input.TicketDescription,
@@ -92,13 +93,19 @@ public class ProjectBoardModel
         _context.Tickets.Add(ticketDetails);
         await _context.SaveChangesAsync();
 
+        var notificationId = _notificationModel.StoreNotification(_context, projectId, ticketDetails.StringId, username, "", ticketDetails.Status, NotificationType.AddTicket);
+
         if (ticketDetails?.Id == null || ticketDetails.Id == 0)
         {
             return new OperationResult { Success = false, Message = "Unable to create new ticket" };
         }
 
+
         projectDetails.TicketsId.Add(ticketDetails.Id);
+        projectDetails.NotificationIds.Add(notificationId);
+        projectDetails.NotificationCount += 1;
         await _context.SaveChangesAsync();
+
         return new OperationResult { Success = true, Message = "Ticket created successfully!" };
     }
 
@@ -119,7 +126,7 @@ public class ProjectBoardModel
 
             _burndownModel.StoreBurndownData(_context, projectId);
 
-            Tickets =  tickets;
+            Tickets = tickets;
         }
         catch (Exception ex)
         {
@@ -173,12 +180,27 @@ public class ProjectBoardModel
         }
 
         ticketDetails.Status = Enum.Parse<TicketStatus>(status);
+        var username = _userManager?.GetUserAsync(user)?.Result?.FirstName + " " + _userManager?.GetUserAsync(user)?.Result?.LastName;
+        if (username == null)
+        {
+            return new OperationResult { Success = false, Message = "User is not authenticated" };
+        }
+
+        var projectDetails = await _context.Projects.FirstOrDefaultAsync(p => p.Id == projectId);
+        if (projectDetails == null)
+        {
+            return new OperationResult { Success = false, Message = "User is not authenticated" };
+        }
 
         try
         {
             await _context.SaveChangesAsync();
 
+            var notificationId = _notificationModel.StoreNotification(_context, projectId, ticketDetails.StringId, username, "", ticketDetails.Status, NotificationType.MoveTicket);
+            projectDetails.NotificationIds.Add(notificationId);
+            projectDetails.NotificationCount += 1;
             _burndownModel.StoreBurndownData(_context, projectId);
+            await _context.SaveChangesAsync();
 
             return new OperationResult { Success = true, Message = $"Status of ticket changed successfully to {status}." };
         }
@@ -203,13 +225,51 @@ public class ProjectBoardModel
             return new OperationResult { Success = false, Message = "User is not authenticated" };
         }
 
-        ticketDetails.Priority = input.TicketPriorityValue;
-        ticketDetails.Status = input.TicketStatusValue;
-        ticketDetails.Description = input.TicketDescription;
-        ticketDetails.Title = input.TicketTitle;
+        bool ticketChange = TicketChanges(ticketDetails, input);
+        bool statusChange = false;
 
+        if (ticketDetails.Status != input.TicketStatusValue)
+        {
+            ticketDetails.Status = input.TicketStatusValue;
+            statusChange = true;
+        }
+
+        if (ticketChange == false && statusChange == false)
+        {
+            return new OperationResult { Success = false, Message = "No changes made to ticket" };
+        }
+
+        var username = _userManager?.GetUserAsync(user)?.Result?.FirstName + " " + _userManager?.GetUserAsync(user)?.Result?.LastName;
+        if (username == null)
+        {
+            return new OperationResult { Success = false, Message = "User is not authenticated" };
+        }
+
+        var projectDetails = await _context.Projects.FirstOrDefaultAsync(p => p.Id == projectId);
+        if (projectDetails == null)
+        {
+            return new OperationResult { Success = false, Message = "User is not authenticated" };
+        }
         try
         {
+            await _context.SaveChangesAsync();
+
+            var notificationId = 0;
+
+            if (ticketChange == true)
+            {
+                notificationId = _notificationModel.StoreNotification(_context, projectId, ticketDetails.StringId, username, "", ticketDetails.Status, NotificationType.EditTicket);
+                projectDetails.NotificationIds.Add(notificationId);
+                projectDetails.NotificationCount += 1;
+            }
+
+            if (statusChange == true)
+            {
+                notificationId = _notificationModel.StoreNotification(_context, projectId, ticketDetails.StringId, username, "", ticketDetails.Status, NotificationType.MoveTicket);
+                projectDetails.NotificationIds.Add(notificationId);
+                projectDetails.NotificationCount += 1;
+            }
+
             await _context.SaveChangesAsync();
 
             _burndownModel.StoreBurndownData(_context, projectId);
@@ -224,6 +284,32 @@ public class ProjectBoardModel
         }
     }
 
+    public bool TicketChanges(Ticket ticket, InputModel input)
+    {
+        bool changed = false;
+
+        if (ticket.Priority != input.TicketPriorityValue)
+        {
+            ticket.Priority = input.TicketPriorityValue;
+            changed = true;
+        }
+
+        if (ticket.Description != input.TicketDescription)
+        {
+            ticket.Description = input.TicketDescription;
+            changed = true;
+        }
+
+        if (ticket.Title != input.TicketTitle)
+        {
+            ticket.Title = input.TicketTitle;
+            changed = true;
+        }
+
+        return changed;
+    }
+
+
     public async Task<OperationResult> DeleteTicket(ClaimsPrincipal user, int projectId, int ticketId)
     {
         if (ValidUserClaim(user, projectId) == false)
@@ -237,10 +323,27 @@ public class ProjectBoardModel
             return new OperationResult { Success = false, Message = "User is not authenticated" };
         }
 
+        var username = _userManager?.GetUserAsync(user)?.Result?.FirstName + " " + _userManager?.GetUserAsync(user)?.Result?.LastName;
+        if (username == null)
+        {
+            return new OperationResult { Success = false, Message = "User is not authenticated" };
+        }
+
+        var projectDetails = await _context.Projects.FirstOrDefaultAsync(p => p.Id == projectId);
+        if (projectDetails == null)
+        {
+            return new OperationResult { Success = false, Message = "User is not authenticated" };
+        }
+
         try
         {
             _context.Tickets.Remove(ticket);
             await _context.SaveChangesAsync();
+            var notificationId = _notificationModel.StoreNotification(_context, projectId, ticket.StringId, username, "", ticket.Status, NotificationType.DeleteTicket);
+            projectDetails.NotificationIds.Add(notificationId);
+            projectDetails.NotificationCount += 1;
+            await _context.SaveChangesAsync();
+
         }
         catch (Exception ex)
         {
